user_tests/requirements.txt
flask
sentence-transformers
qdrant-client
psycopg2-binary
boto3
numpy
torch
